+++ include/managers/settings_manager.h
@@ -19,7 +19,8 @@ typedef enum {
 // idle/status oled animation mode
 typedef enum {
   IDLE_ANIM_GAME_OF_LIFE = 0,
-  IDLE_ANIM_GHOST = 1
+  IDLE_ANIM_GHOST = 1,
+  IDLE_ANIM_STARFIELD = 2
 } IdleAnimation;
 #endif